jpa logging config update and test repo add

+++ src/main/kotlin/kr/co/osj4532/model/entity/TestMst.kt
@@ -1,14 +1,12 @@
 package kr.co.osj4532.model.entity
 
 import javax.persistence.Entity
-import javax.persistence.GeneratedValue
-import javax.persistence.GenerationType
 import javax.persistence.Id
 
 @Entity(name = "test")
 data class TestMst (
         @Id
         // GenerationType.IDENTITY 쓰기 지연 X
-        val id: Int = 0,
+        val id: Long = 0,
         val name: String? = null
 )
